feat: extract meeting dialogs from __main__ into meeting_dialogs.py
Hardening item 6, extraction 2a (architecture spec A2). The Tk dialog
builders move to MeetingDialogs: ask_meeting_name,
ask_speaker_confirmation (incl. the deep-ID rerun), ask_recovery_name,
show_llm_unavailable. Styling helpers (style_window, flat_button,
center_window, run_modal, sanitize_name) and the ABORT sentinel become
module level so the remaining flows share them. meeting_job routes
through app.dialogs. Direct move; layouts, outcome logging, and the
abort-on-crash contract unchanged. __main__.py drops ~730 lines.

9 new tests pin sanitize_name and the abort-on-crash contract for
every dialog. System suite 223 passed.
