+++ CLAUDE.md
@@ -22,6 +22,7 @@ When a user reports that a meeting "crashed" or "failed" or "didn't finish", do 
 |--------|---------|
 | `__main__.py` | Entry point. Tray icon, hotkey wiring, meeting flow, menu/settings, GitHub status, session stats. Workflows are being extracted into flow modules (hardening item 6). |
 | `dictation_flow.py` | Dictation workflow component: normal/feature-suggest/overlay dictation, auto-stop cap, discard, crash recovery, history. Feature routing state lives in AppState.feature_suggest. |
+| `meeting_dialogs.py` | Meeting Tk dialogs (name, speaker confirmation, recovery, LLM-unavailable) + shared styling helpers and the ABORT sentinel. All run on the DialogDispatcher thread. |
 | `state_manager.py` | Observable state machine. AppState dataclass, typed StateEvent, event constants, thread-safe emit(), listener subscriptions (on/on_any), event log ringbuffer. All icon/toast updates flow through here. |
 | `transcribe.py` | WhisperX with persistent model cache. Fast path (dictation) and staged pipeline (meeting) |
 | `worker.py` | Multiprocessing transcription worker. CUDA isolation via spawn context |
