+++ whisper_sync/meeting_job.py
@@ -263,7 +263,7 @@ class MeetingJob:
             if id_result and id_result.get("speaker_map"):
                 try:
                     self.app._current_meeting_json_path = json_path
-                    confirmation = self.app._ask_speaker_confirmation(id_result)
+                    confirmation = self.app.dialogs.ask_speaker_confirmation(id_result)
                     if confirmation:
                         if isinstance(confirmation, tuple):
                             confirmed_map, boundaries = confirmation
